+++ example/lib/counter/counter_cube_widget_animation.dart
@@ -36,7 +36,7 @@ class CounterCubeWidgetAnimation extends CubeWidgetAnimation<CounterCube> {
           child: Column(
             mainAxisAlignment: MainAxisAlignment.center,
             children: <Widget>[
-              Text(getString('description_counter')),
+              Text(Cubes.getString('description_counter')),
               cube.count.build<int>((value) {
                 return Text(value.toString());
               }),