+++ lib/src/widgets/cube_widget.dart
@@ -18,11 +18,11 @@ abstract class CubeWidget<C extends Cube> extends StatelessWidget {
   Widget build(BuildContext context) {
     return CubeBuilder<C>(
       builder: buildView,
+      initData: initData,
       onError: (cube, text) => onError(context, cube, text),
       onSuccess: (cube, text) => onSuccess(context, cube, text),
       onAction: (cube, data) => onAction(context, cube, data),
       initState: (cube) => initState(context, cube),
-      initData: initData,
       dispose: dispose,
     );
   }