+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+# 0.4.0
+
+- [BREAKING CHANGE] move `getString` to `Cubes.getString`
+- [BREAKING CHANGE] move `registerDependency` to `Cubes.registerDependency`
+- [BREAKING CHANGE] move `getDependency` to `Cubes.getDependency`
+- [BREAKING CHANGE] remove `getCube` (use getDependency)
+- [BREAKING CHANGE] remove `registerCube` (use registerDependency)
+- create `Cubes.instance.customInjector(Injector)` to use if you want to overwrite the default injector (getIt).
+- add test widget example of the CounterScreen.
+
 # 0.3.5
 
 - improvements in `CubeWidgetAnimation` [still experimental]